requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0 